requirements.txt
flask-wtf
email-validator == 1.0.5
flask-sqlalchemy
flask-bcrypt
flask-login
Pillow
elasticsearch
elasticsearch_dsl